requirements.txt
fastapi
pydantic
uvicorn
python-dotenv
requests
openai
